refactor(#1868): PR 22 — dispatch.rs fully pure Rust, extract hook_registry.rs
Extract all PyO3-dependent types from dispatch.rs to new hook_registry.rs:
- InterceptHook trait (methods use Py<PyAny>/PyErr)
- HookEntry + HookRegistry (stores Box<dyn InterceptHook> + Py<PyAny>)
- ObserverEntry + ObserverRegistry (stores Py<PyAny>)

dispatch.rs is now 100% pure Rust (zero PyO3 imports):
- PathResolver trait (pure Rust)
- MutationObserver trait (pure Rust)
- PathTrie (pure Rust, ~50ns lookup)

Pure Rust kernel files (zero PyO3):
  kernel.rs, dispatch.rs, router.rs, backend.rs, metastore.rs

Codegen updated to parse hook_registry.rs for trait definitions.
2551 tests passed, 0 failed.
