+++ rust/nexus_kernel/src/dispatch.rs
@@ -1,58 +1,18 @@
-//! KernelDispatch — dispatch traits, trie, hook/observer registries.
+//! KernelDispatch — pure Rust dispatch traits + PathTrie.
 //!
-//! Three dispatch trait families (PR 7c, hand-written -- PR 8 codegen replaces):
-//!   - InterceptHook: pre/post hooks per syscall (INTERCEPT phase, LSM-style)
+//! Zero PyO3 dependency. All Python-dependent types (InterceptHook trait,
+//! HookRegistry, ObserverRegistry) live in hook_registry.rs.
+//!
+//! Contains:
 //!   - PathResolver: virtual path short-circuit (PRE-DISPATCH phase, procfs-style)
 //!   - MutationObserver: fire-and-forget event notification (OBSERVE phase, fsnotify-style)
+//!   - PathTrie: O(path_depth) lookup (~50ns) for virtual path resolvers
 //!
-//! Infrastructure:
-//!   - PathTrie: O(path_depth) lookup (~50ns) for virtual path resolvers.
-//!   - HookRegistry: cached metadata for INTERCEPT hooks.
-//!   - ObserverRegistry: bitmask-filtered OBSERVE observers.
-//!
-//! All types are owned directly by Kernel (no Arc wrapper, no #[pyclass]).
-//! PyO3 types (Py<PyAny>) are used only in HookEntry/ObserverEntry for storing
-//! original Python objects returned to Python callers — these are opaque handles.
-//!
-//! Issue #1868: Kernel owns all dispatch state.
+//! Issue #1868: PR 22 — dispatch.rs fully pure Rust.
 
 use parking_lot::RwLock;
-use pyo3::prelude::*;
-use pyo3::types::PyAny;
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicUsize, Ordering};
-
-// ── Dispatch Traits (Rust ABI — PR 8 codegen replaces hand-written adapters) ──
-
-/// INTERCEPT hook — called before/after each syscall.
-///
-/// Rust equivalent of Python `VFSReadHook`/`VFSWriteHook`/etc.
-/// Pre-hooks can abort by returning Err. Post-hooks are fire-and-forget.
-///
-/// Each method receives opaque context (PyObject) — kernel never inspects it.
-/// PR 8 codegen will generate per-syscall typed context structs.
-#[allow(dead_code)]
-pub(crate) trait InterceptHook: Send + Sync {
-    fn name(&self) -> &str;
-    fn on_pre_read(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_read(&self, ctx: &Py<PyAny>);
-    fn on_pre_write(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_write(&self, ctx: &Py<PyAny>);
-    fn on_pre_delete(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_delete(&self, ctx: &Py<PyAny>);
-    fn on_pre_rename(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_rename(&self, ctx: &Py<PyAny>);
-    fn on_pre_mkdir(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_mkdir(&self, ctx: &Py<PyAny>);
-    fn on_pre_rmdir(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_rmdir(&self, ctx: &Py<PyAny>);
-    fn on_pre_copy(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_copy(&self, ctx: &Py<PyAny>);
-    fn on_pre_stat(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_stat(&self, ctx: &Py<PyAny>);
-    fn on_pre_access(&self, ctx: &Py<PyAny>) -> Result<(), PyErr>;
-    fn on_post_access(&self, ctx: &Py<PyAny>);
-}
 
 /// PRE-DISPATCH resolver — virtual path short-circuit.
 ///
@@ -233,223 +193,7 @@ impl Trie {
     }
 }
 
-// ── HookRegistry (owned by PyKernel wrapper, not by pure Rust Kernel) ──
-
-/// Cached metadata for a single hook.
-pub(crate) struct HookEntry {
-    /// Rust trait object — used by kernel dispatch (language-agnostic).
-    pub(crate) hook: Box<dyn InterceptHook>,
-    /// Original Python object — returned to Python callers via get_pre_hooks().
-    pub(crate) hook_py: Py<PyAny>,
-    pub(crate) has_pre: bool,
-    pub(crate) is_async_post: bool,
-    #[allow(dead_code)]
-    pub(crate) name: String,
-}
-
-/// Registry that caches hook metadata at registration time.
-///
-/// Eliminates per-dispatch `getattr()` and `inspect.iscoroutinefunction()`
-/// overhead by detecting these properties once at `register()` time.
-pub(crate) struct HookRegistry {
-    ops: HashMap<String, Vec<HookEntry>>,
-}
-
-impl HookRegistry {
-    pub(crate) fn new() -> Self {
-        Self {
-            ops: HashMap::new(),
-        }
-    }
-
-    /// Register a hook for the given operation.
-    pub(crate) fn register(
-        &mut self,
-        op: &str,
-        hook_impl: Box<dyn InterceptHook>,
-        hook_py: Py<PyAny>,
-        has_pre: bool,
-        is_async_post: bool,
-        name: String,
-    ) {
-        self.ops.entry(op.to_string()).or_default().push(HookEntry {
-            hook: hook_impl,
-            hook_py,
-            has_pre,
-            is_async_post,
-            name,
-        });
-    }
-
-    /// Remove a hook by identity (`is` check on original Python object).
-    pub(crate) fn unregister(&mut self, py: Python<'_>, op: &str, hook: &Bound<'_, PyAny>) -> bool {
-        if let Some(entries) = self.ops.get_mut(op) {
-            let hook_ptr = hook.as_ptr();
-            if let Some(pos) = entries
-                .iter()
-                .position(|e| e.hook_py.bind(py).as_ptr() == hook_ptr)
-            {
-                entries.remove(pos);
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Return Python hook objects that have `on_pre_{op}` (for Python callers).
-    pub(crate) fn get_pre_hooks(&self, py: Python<'_>, op: &str) -> Vec<Py<PyAny>> {
-        self.ops
-            .get(op)
-            .map(|entries| {
-                entries
-                    .iter()
-                    .filter(|e| e.has_pre)
-                    .map(|e| e.hook_py.clone_ref(py))
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    /// Return Rust trait references for pre-hooks (for kernel dispatch).
-    pub(crate) fn get_pre_hook_impls(&self, op: &str) -> Vec<&dyn InterceptHook> {
-        self.ops
-            .get(op)
-            .map(|entries| {
-                entries
-                    .iter()
-                    .filter(|e| e.has_pre)
-                    .map(|e| e.hook.as_ref())
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    /// Return Rust trait references for sync post-hooks (kernel dispatch).
-    pub(crate) fn get_post_hook_impls(&self, op: &str) -> Vec<&dyn InterceptHook> {
-        self.ops
-            .get(op)
-            .map(|entries| {
-                entries
-                    .iter()
-                    .filter(|e| !e.is_async_post)
-                    .map(|e| e.hook.as_ref())
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-
-    /// Return (sync_post_hooks, async_post_hooks) as Python objects.
-    pub(crate) fn get_post_hooks(
-        &self,
-        py: Python<'_>,
-        op: &str,
-    ) -> (Vec<Py<PyAny>>, Vec<Py<PyAny>>) {
-        let entries = match self.ops.get(op) {
-            Some(e) => e,
-            None => return (Vec::new(), Vec::new()),
-        };
-        let sync: Vec<Py<PyAny>> = entries
-            .iter()
-            .filter(|e| !e.is_async_post)
-            .map(|e| e.hook_py.clone_ref(py))
-            .collect();
-        let async_: Vec<Py<PyAny>> = entries
-            .iter()
-            .filter(|e| e.is_async_post)
-            .map(|e| e.hook_py.clone_ref(py))
-            .collect();
-        (sync, async_)
-    }
-
-    /// Return all Python hook objects for the given operation.
-    pub(crate) fn get_all_hooks(&self, py: Python<'_>, op: &str) -> Vec<Py<PyAny>> {
-        self.ops
-            .get(op)
-            .map(|entries| entries.iter().map(|e| e.hook_py.clone_ref(py)).collect())
-            .unwrap_or_default()
-    }
-
-    /// Number of hooks registered for the given operation.
-    pub(crate) fn count(&self, op: &str) -> usize {
-        self.ops.get(op).map(|e| e.len()).unwrap_or(0)
-    }
-}
-
-// ── ObserverRegistry (owned by PyKernel wrapper) ────────────────────
-
-pub(crate) struct ObserverEntry {
-    pub(crate) observer: Py<PyAny>,
-    pub(crate) name: String,
-    pub(crate) event_mask: u32,
-}
-
-/// Rust-side observer registry with event-type bitmask filtering.
-pub(crate) struct ObserverRegistry {
-    observers: Vec<ObserverEntry>,
-}
-
-impl ObserverRegistry {
-    pub(crate) fn new() -> Self {
-        Self {
-            observers: Vec::new(),
-        }
-    }
-
-    /// Register observer with event_mask bitmask.
-    pub(crate) fn register(
-        &mut self,
-        py: Python<'_>,
-        obs: Py<PyAny>,
-        event_mask: u32,
-    ) -> PyResult<()> {
-        let obs_ref = obs.bind(py);
-        let name: String = obs_ref
-            .get_type()
-            .name()
-            .map(|n| n.to_string())
-            .unwrap_or_else(|_| "<?>".to_string());
-
-        self.observers.push(ObserverEntry {
-            observer: obs,
-            name,
-            event_mask,
-        });
-        Ok(())
-    }
-
-    /// Unregister by identity.
-    pub(crate) fn unregister(&mut self, py: Python<'_>, obs: &Bound<'_, PyAny>) -> bool {
-        let obs_ptr = obs.as_ptr();
-        if let Some(pos) = self
-            .observers
-            .iter()
-            .position(|e| e.observer.bind(py).as_ptr() == obs_ptr)
-        {
-            self.observers.remove(pos);
-            return true;
-        }
-        false
-    }
-
-    /// Return (observer, name) pairs matching the event_type_bit.
-    pub(crate) fn get_matching(
-        &self,
-        py: Python<'_>,
-        event_type_bit: u32,
-    ) -> Vec<(Py<PyAny>, String)> {
-        self.observers
-            .iter()
-            .filter(|e| e.event_mask & event_type_bit != 0)
-            .map(|e| (e.observer.clone_ref(py), e.name.clone()))
-            .collect()
-    }
-
-    pub(crate) fn count(&self) -> usize {
-        self.observers.len()
-    }
-}
-
-// ── Tests (TrieNode only — no PyO3 linking required) ───────────────────
+// ── Tests (TrieNode only — pure Rust, no PyO3) ────────────────────────
 
 #[cfg(test)]
 mod tests {
