+++ src/nexus/core/kernel_protocols.py
@@ -15,6 +15,18 @@ from typing import Any, Protocol, runtime_checkable
 
 
 @runtime_checkable
+class PathResolver(Protocol):
+    def try_read(self, path: str) -> bytes | None: ...
+    def try_write(self, path: str, content: bytes) -> None | None: ...
+    def try_delete(self, path: str) -> None | None: ...
+
+
+@runtime_checkable
+class MutationObserver(Protocol):
+    def on_mutation(self, event_type: int, path: str) -> None: ...
+
+
+@runtime_checkable
 class InterceptHook(Protocol):
     def name(self) -> str: ...
     def on_pre_read(self, ctx: Any) -> Any: ...
@@ -35,18 +47,6 @@ class InterceptHook(Protocol):
     def on_post_stat(self, ctx: Any) -> None: ...
     def on_pre_access(self, ctx: Any) -> Any: ...
     def on_post_access(self, ctx: Any) -> None: ...
-
-
-@runtime_checkable
-class PathResolver(Protocol):
-    def try_read(self, path: str) -> bytes | None: ...
-    def try_write(self, path: str, content: bytes) -> None | None: ...
-    def try_delete(self, path: str) -> None | None: ...
-
-
-@runtime_checkable
-class MutationObserver(Protocol):
-    def on_mutation(self, event_type: int, path: str) -> None: ...
 
 
 @runtime_checkable
